# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "Petship API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://api.petship.pt",
            "description": "Petship API"
        }
    ],
    "paths": {
...
        "/api/novoanuncio": {
            "post": {
                "tags": [
                    "Anuncios"
                ],
                "summary": "Criar um anuncio",
                "description": "Rota para criar um anuncio, o utilizador tem de estar logado. Se o anuncio for criado com sucesso retorna o status 200",
                "operationId": "1dce6c0f5c7034abc822fc089714c2ef",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "sexo",
                                    "especie",
                                    "raca",
                                    "porte",
                                    "idade",
                                    "cor",
                                    "distrito",
                                    "etiqueta"
                                ],
                                "properties": {
                                    "nome": {
                                        "example": "Rufus"
                                    },
                                    "sexo": {
                                        "example": "1"
                                    },
                                    "especie": {
                                        "example": "1"
                                    },
                                    "raca": {
                                        "example": "1"
                                    },
                                    "porte": {
                                        "example": "1"
                                    },
                                    "idade": {
                                        "example": "1"
                                    },
                                    "cor": {
                                        "example": "1"
                                    },
                                    "distrito": {
                                        "example": "Coimbra"
                                    },
                                    "etiqueta": {
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/animal/{id}": {
            "get": {
                "tags": [
                    "Anuncios"
                ],
                "summary": "Listar animais para adoção",
                "description": "",
                "operationId": "6eeaaf2a04a3594255e071bc5d082cf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do animal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/utilizador/anuncios": {
            "get": {
                "tags": [
                    "Anuncios"
                ],
                "summary": "Listar anuncios de um utilizador",
                "description": "",
                "operationId": "d676196acdb8b5fc3ea6d6c463d0df1c",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/removeranuncio/{id}": {
            "get": {
                "tags": [
                    "Anuncios"
                ],
                "summary": "Remover anuncio",
                "description": "",
                "operationId": "8fa18fb01189ad19f502842b39065aa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do anuncio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/anuncio/num/{id}": {
            "get": {
                "tags": [
                    "Anuncios"
                ],
                "summary": "Dados de um anuncio (numéricos)",
                "description": "",
                "operationId": "5b1ee0564bfab367a69344bd66ae5d32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do anuncio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/editaranuncio/{id}": {
            "post": {
                "tags": [
                    "Anuncios"
                ],
                "summary": "Editar anuncio",
                "description": "",
                "operationId": "d4029006a3097e865b0d616ab469fd26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do anuncio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/anuncio/estado/{id}": {
            "get": {
                "tags": [
                    "Anuncios"
                ],
                "summary": "Alterar estado do anuncio",
                "description": "",
                "operationId": "b6de613dc0c2119ed63a0398ca52149d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do anuncio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}